rename to icetrayless and add a few more query result funcs
