+++ skymap_scanner/utils/icetrayless.py
